test: share vault file adapter fixture
