+++ tests/unit/providers/claude/storage/SkillStorage.test.ts
@@ -1,39 +1,5 @@
-import type { VaultFileAdapter } from '@/core/storage/VaultFileAdapter';
+import { createMockVaultFileAdapter as createMockAdapter } from '@test/helpers/MockVaultFileAdapter';
 import { SKILLS_PATH,SkillStorage } from '@/providers/claude/storage/SkillStorage';
-
-function createMockAdapter(files: Record<string, string> = {}): VaultFileAdapter {
-  const mockAdapter = {
-    exists: jest.fn(async (path: string) => path in files || Object.keys(files).some(k => k.startsWith(path + '/'))),
-    read: jest.fn(async (path: string) => {
-      if (!(path in files)) throw new Error(`File not found: ${path}`);
-      return files[path];
-    }),
-    write: jest.fn(),
-    delete: jest.fn(),
-    listFolders: jest.fn(async (folder: string) => {
-      const prefix = folder.endsWith('/') ? folder : folder + '/';
-      const folders = new Set<string>();
-      for (const path of Object.keys(files)) {
-        if (path.startsWith(prefix)) {
-          const rest = path.slice(prefix.length);
-          const firstSlash = rest.indexOf('/');
-          if (firstSlash >= 0) {
-            folders.add(prefix + rest.slice(0, firstSlash));
-          }
-        }
-      }
-      return Array.from(folders);
-    }),
-    listFiles: jest.fn(),
-    listFilesRecursive: jest.fn(),
-    ensureFolder: jest.fn(),
-    rename: jest.fn(),
-    append: jest.fn(),
-    stat: jest.fn(),
-    deleteFolder: jest.fn(),
-  } as unknown as VaultFileAdapter;
-  return mockAdapter;
-}
 
 describe('SkillStorage', () => {
   it('exports SKILLS_PATH', () => {
