+++ tests/unit/providers/codex/storage/CodexSubagentStorage.test.ts
@@ -1,6 +1,6 @@
 import { TEST_CODEX_MODEL } from '@test/helpers/codexModels';
+import { createMockVaultFileAdapter as createMockAdapter } from '@test/helpers/MockVaultFileAdapter';
 
-import type { VaultFileAdapter } from '@/core/storage/VaultFileAdapter';
 import {
   CODEX_AGENTS_PATH,
   CodexSubagentStorage,
@@ -10,33 +10,6 @@ import {
   serializeSubagentToml,
 } from '@/providers/codex/storage/CodexSubagentStorage';
 import type { CodexSubagentDefinition } from '@/providers/codex/types/subagent';
-
-function createMockAdapter(files: Record<string, string> = {}): VaultFileAdapter {
-  return {
-    exists: jest.fn(async (path: string) =>
-      path in files || Object.keys(files).some(k => k.startsWith(path + '/')),
-    ),
-    read: jest.fn(async (path: string) => {
-      if (!(path in files)) throw new Error(`File not found: ${path}`);
-      return files[path];
-    }),
-    write: jest.fn(),
-    delete: jest.fn(),
-    listFiles: jest.fn(async (folder: string) => {
-      const prefix = folder.endsWith('/') ? folder : folder + '/';
-      return Object.keys(files).filter(
-        k => k.startsWith(prefix) && !k.slice(prefix.length).includes('/'),
-      );
-    }),
-    listFolders: jest.fn(),
-    listFilesRecursive: jest.fn(),
-    ensureFolder: jest.fn(),
-    rename: jest.fn(),
-    append: jest.fn(),
-    stat: jest.fn(),
-    deleteFolder: jest.fn(),
-  } as unknown as VaultFileAdapter;
-}
 
 const BASIC_TOML = `name = "reviewer"
 description = "PR reviewer focused on correctness."
